Created a survey result page that doesn't require password

+++ src/main/java/com/surveyMonkey/util/StoreAnswerHelper.java
@@ -4,7 +4,7 @@ public class StoreAnswerHelper {
     private long questionId;
     private String answer;
 
-    public void setQuestionId(int questionId) {
+    public void setQuestionId(long questionId) {
         this.questionId = questionId;
     }
